added a helper class
its actually better in general to do the opposite for this kind fo work. you should first build the helper class and ask students to fill in methods there. then put calls to that helper class in the original. in this way the original structure of the project stays the same, and all you to do is pull out logic from whatever file.